Strip the CAS ticket from the service parameter when validating tickets.
Without this change, the first attempt at validating the ticket will fail
because the service-url is different from the URL passed when the user is
first redirected to the CAS server:

  <cas:serviceResponse xmlns:cas='http://www.yale.edu/tp/cas'>
    <cas:authenticationFailure code='INVALID_SERVICE'>
      ticket &#039;ST-98762-ev9L0fXj2cGipqda324f&#039; does not
      match supplied service.  The original service was
      &#039;http://localhost:8000/&#039; and the supplied service was
      &#039;http://localhost:8000/?ticket=ST-98762-ev9L0fXj2cGipqda324f&#039;
    </cas:authenticationFailure>\n
  </cas:serviceResponse>

Authentication would still be successful, but the user would be redirected
back to the CAS host with a ticket parameter in the URL and would have a
*second* ticket generated for the application which would then validate, I
think because `$request->get($this->query_ticket_parameter)` would only
return the second ticket when two are present, behavior that shouldn't be
relied upon.

By stripping the ticket from the service URL, the first ticket can be properly
validated without an additional back-and-forth.